Make local badge persistent (EJABS-2114)
When the device is offline,  we must sent as badge number:
number of messages in spool + local badge count for the device.

For that, we need to store the local badge count informed.

There is a new column in the sql applepush_cache table : 'local_badge'.

Mnesia table isn't changed, we store it as part of the the key-value options