Handle disable push stanzas on plan sessions (EJSAAS-203)
Client can specify the specific appid and token it want to remove
from pushes:

<iq type='set' id='123'>
 <disable xmlns='p1:push'>
   <appid>APP_ID</appid>
    <notification>
       <type>gcm</type>
       <id>TOKEN</id>
    </notification>
 </disable>
</iq>

If the session has push enabled,  and the client send the appid-token
when disabling,  it must match with the one in session.
If the session has no push enabled, and the client doesn't specify
the token when disabling,  it is an error condition and the appropiate
error reply is sent.

If the session has no push enabled, but the client specify the app
and deviceid when disabling notifications,  that token is deleted
from cache,  and offline pushes to that device are disabled.  This
allow clients to login and disable offline pushes, without first
enabling it on the session.