Race condition on session replacement of detached sessions (EJABS-2443)
When sessions are in detached state, they can have queued messages on
it.
If they are replaced (full login using the same resource, not using rebind),
the queued messages must be delivered to the new session.  The way this was
done introduce race conditions on between the new session registration
on SM,  the old session re-routing messages,  and offline storage.
See descriptions on https://support.process-one.net/browse/EJABS-2443

This try to avoid that by passing the PID of the new session to the one
being replaced, letting it pass the messages directly.  This is very
similar to how rebind works.

Not sure this handles all the cases, but should be an improvement over
the current behaviour.